TST: Deleted old test helper modules
Most of the code is deleted, some moved to appropriate
test modules
